requirements.txt
streamlit
python-dotenv
requests
fpdf
python-docx
yt-dlp
lxml